fix lint for app name
