+++ client/src/domains/core/app-name/APP_NAME.ts
@@ -1,0 +1,5 @@
+import { InjectionToken } from '@angular/core';
+
+export const APP_NAME = new InjectionToken<string>(
+  'name of the application for display prupose',
+);
