+++ client/src/domains/core/app-name/provideAppName.ts
@@ -1,0 +1,11 @@
+import { EnvironmentProviders, makeEnvironmentProviders } from '@angular/core';
+import { APP_NAME } from './APP_NAME';
+
+export function provideAppName(): EnvironmentProviders {
+  return makeEnvironmentProviders([
+    {
+      provide: APP_NAME,
+      useValue: 'Geschmack nach Christlicher Art',
+    },
+  ]);
+}
