+++ client/src/domains/core/index.ts
@@ -1,4 +1,5 @@
 export * from './core.routes';
 export * from './app-title-strategy/provideAppTitleStrategy';
 export * from './layout/index';
-export * from './provideAppName';
+export * from './app-name/APP_NAME';
+export * from './app-name/provideAppName';
